+++ crates/bones3_core/src/util/mod.rs
@@ -1,0 +1,4 @@
+//! This module contains useful utility components and systems that might be
+//! used often while working with Bones Cubed.
+
+pub mod anchor;