+++ crates/bones3_remesh/src/query/mod.rs
@@ -1,0 +1,6 @@
+//! This module contains extensions and overrides for VoxelCommands and
+//! VoxelQueries that are useful for generating chunk meshes.
+
+mod commands;
+
+pub use commands::*;