+++ src/lib.rs
@@ -10,11 +10,10 @@
 pub use bones3_core as core;
 #[cfg(feature = "meshing")]
 pub use bones3_remesh as remesh;
+#[cfg(feature = "worldgen")]
+pub use bones3_worldgen as worldgen;
 
 /// Used to import common components and systems for Bones Cubed.
-#[allow(ambiguous_glob_reexports)]
 pub mod prelude {
     pub use super::core::prelude::*;
-    #[cfg(feature = "meshing")]
-    pub use super::remesh::prelude::*;
 }